requirements.txt
Flask==1.1.2
Pillow==8.3.2
tensorflow==2.5.0